Start on a 'hooks' system
This will allow scripts to specify
a 'handler' to get called when
certain things happen. It's a bit
like an event listener.

FOR USERS: Enjoy the scripts you use getting more powerful.

FOR SCRIPTERS: Use `bp.hooks.addToHook()` to register
a custom handler for a hook. Two parameters are required:
`hook`, which is the hook ID that you want to attatch your handler to,
and `handler`, which is a javascript function to be executed
when the hook is called. You can atatch multiple handlers to a hook.
Note that creating new hooks is internal-only: scripts can't
define their own hooks. That would be way too complicated.
TODO: A way to remove hook handlers? Namespaced hook handlers?

FOR BOATPAD DEVS: To create a new hook, simply add it to the directory
at `bp.hooks.list`. Try to keep hook namespaces consistent with command
namespaces. Specify the point in the code where a hook is called using
runHook(). It takes one parameter: `hook`, which is the ID for the hook
you want to call. Try to avoid calling the ame hook twice. Maybe it will
be appropiate in some cases. I dunnno.

P.S. I also fixed a bug at line 122